llm-training-platform/src/document_ingestion/requirements.txt
# Document Ingestion Service Requirements

# OCR and Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
python-docx>=0.8.11
pdfminer.six>=20221105
pillow>=9.5.0
opencv-python>=4.7.0.72
langdetect>=1.0.9
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# FastAPI and Web
fastapi>=0.95.1
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Database and Storage
sqlalchemy>=2.0.12
psycopg2-binary>=2.9.6
alembic>=1.10.4

# Utilities
loguru>=0.7.0
pydantic>=1.10.7
python-dotenv>=1.0.0
tenacity>=8.2.2
